added active Filter Badge Component
