+++ packages/admin/src/table/filterbar/index.ts
@@ -1,2 +1,3 @@
 export * from "./FilterBar";
 export * from "./FilterBarPopOverFormField";
+export * from "./FilterBarActiveFilterBadge";
